Revert "Viewport: keep clip height in small z"
This reverts commit e8e0b7df69028437da2016ee44d3c04cb49f25a4.

+++ src/OpenLoco/include/OpenLoco/Viewport.hpp
@@ -250,7 +250,7 @@ namespace OpenLoco::Ui
         }
     }
 
-    uint8_t getMaxClipHeight();
-    void setMaxClipHeight(uint8_t maxClipHeight);
+    int16_t getMaxClipHeight();
+    void setMaxClipHeight(int16_t maxClipHeight);
     void resetMaxClipHeight();
 }
